requirement.txt
pandas==1.1.4Django==3.1.3djangorestframework==3.12.2drf-yasg==1.20.0pytorch==1.6.0torchvision==0.7.0cudatoolkit=10.1pororo